requirements.txt
# python packages needed

pandas
numpy
matplotlib
seaborn

# supporting libraries
python-dotenv
rich
streamlit
yfinance
transformers
nltk
requests
Pillow

black
psycopg2


# LLM APIs
openai
anthropic
google-generativeai
groq
duckduckgo-search
exa_py

# llama_index LLM support
llama-index
llama-index-llms-gemini
llama-index-llms-anthropic
llama-index-llms-groq
llama-index-llms-ollama
llama-index-embeddings-huggingface
# for Jina embeddings
llama-index-embeddings-jinaai
# for reading GMail
llama-index-readers-google

# Phidata
phidata



